+++ projectroles/tests/test_remote_project_api.py
@@ -824,7 +824,7 @@ class SyncRemoteDataTestBase(
         self.remote_api = RemoteProjectAPI()
 
         # Default data to receive from source when testing site in target mode
-        self.default_data = {
+        self.remote_data = {
             'users': {
                 SOURCE_USER_UUID: {
                     'sodar_uuid': SOURCE_USER_UUID,
@@ -916,15 +916,9 @@ class SyncRemoteDataTestBase(
 class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
     """Tests for project creation with sync_remote_data()"""
 
-    def test_create(self):
-        """Test sync with non-existing project data and READ_ROLE access"""
-        self.assertEqual(Project.objects.all().count(), 0)
-        self.assertEqual(RoleAssignment.objects.all().count(), 0)
-        self.assertEqual(User.objects.all().count(), 1)
-        self.assertEqual(RemoteProject.objects.all().count(), 0)
-        self.assertEqual(RemoteSite.objects.all().count(), 1)
-
-        self.default_data['users'].update(
+    def _update_remote_data(self):
+        """Update user and role data for remote creation"""
+        self.remote_data['users'].update(
             {
                 SOURCE_USER2_UUID: {
                     'sodar_uuid': SOURCE_USER2_UUID,
@@ -958,7 +952,7 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
                 },
             }
         )
-        self.default_data['projects'][SOURCE_CATEGORY_UUID]['roles'].update(
+        self.remote_data['projects'][SOURCE_CATEGORY_UUID]['roles'].update(
             {
                 SOURCE_CATEGORY_ROLE2_UUID: {
                     'user': SOURCE_USER2_USERNAME,
@@ -974,7 +968,7 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
                 },
             }
         )
-        self.default_data['projects'][SOURCE_PROJECT_UUID]['roles'].update(
+        self.remote_data['projects'][SOURCE_PROJECT_UUID]['roles'].update(
             {
                 SOURCE_PROJECT_ROLE2_UUID: {
                     'user': SOURCE_USER2_USERNAME,
@@ -990,11 +984,19 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
                 },
             }
         )
-        og_data = deepcopy(self.default_data)
-        # Do sync
-        self.remote_api.sync_remote_data(self.source_site, self.default_data)
 
-        # Assert database status
+    def test_create_assert_db(self):
+        """Test sync for creation and assert database changes"""
+        self.assertEqual(Project.objects.all().count(), 0)
+        self.assertEqual(RoleAssignment.objects.all().count(), 0)
+        self.assertEqual(User.objects.all().count(), 1)
+        self.assertEqual(RemoteProject.objects.all().count(), 0)
+        self.assertEqual(RemoteSite.objects.all().count(), 1)
+        self.assertEqual(AppSetting.objects.count(), 0)
+
+        self._update_remote_data()
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
+
         self.assertEqual(Project.objects.all().count(), 2)
         self.assertEqual(RoleAssignment.objects.all().count(), 8)
         self.assertEqual(User.objects.all().count(), 5)
@@ -1002,10 +1004,13 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         self.assertEqual(RemoteSite.objects.all().count(), 2)
         self.assertEqual(AppSetting.objects.count(), 3)
 
-        new_user = User.objects.get(username=SOURCE_USER_USERNAME)
-        new_user2 = User.objects.get(username=SOURCE_USER2_USERNAME)
-        new_user3 = User.objects.get(username=SOURCE_USER3_USERNAME)
-        new_user4 = User.objects.get(username=SOURCE_USER4_USERNAME)
+    def test_create_assert_projects(self):
+        """Test sync for creation and assert project changes"""
+        self.assertEqual(Project.objects.all().count(), 0)
+        self._update_remote_data()
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
+        self.assertEqual(Project.objects.all().count(), 2)
+
         category_obj = Project.objects.get(sodar_uuid=SOURCE_CATEGORY_UUID)
         expected = {
             'id': category_obj.pk,
@@ -1024,6 +1029,35 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         model_dict.pop('readme', None)
         self.assertEqual(model_dict, expected)
 
+        project_obj = Project.objects.get(sodar_uuid=SOURCE_PROJECT_UUID)
+        expected = {
+            'id': project_obj.pk,
+            'title': SOURCE_PROJECT_TITLE,
+            'type': PROJECT_TYPE_PROJECT,
+            'description': SOURCE_PROJECT_DESCRIPTION,
+            'parent': category_obj.pk,
+            'public_access': None,
+            'public_guest_access': False,  # DEPRECATED
+            'archive': False,
+            'full_title': SOURCE_PROJECT_FULL_TITLE,
+            'has_public_children': False,
+            'sodar_uuid': uuid.UUID(SOURCE_PROJECT_UUID),
+        }
+        model_dict = model_to_dict(project_obj)
+        model_dict.pop('readme', None)
+        self.assertEqual(model_dict, expected)
+
+    def test_create_assert_category_roles(self):
+        """Test sync for creation and assert category role changes"""
+        self.assertEqual(RoleAssignment.objects.all().count(), 0)
+        self._update_remote_data()
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
+        self.assertEqual(RoleAssignment.objects.all().count(), 8)
+        new_user = User.objects.get(username=SOURCE_USER_USERNAME)
+        new_user2 = User.objects.get(username=SOURCE_USER2_USERNAME)
+        new_user3 = User.objects.get(username=SOURCE_USER3_USERNAME)
+        new_user4 = User.objects.get(username=SOURCE_USER4_USERNAME)
+        category_obj = Project.objects.get(sodar_uuid=SOURCE_CATEGORY_UUID)
         c_owner_obj = RoleAssignment.objects.get(
             sodar_uuid=SOURCE_CATEGORY_ROLE_UUID
         )
@@ -1036,7 +1070,6 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         c_guest_obj = RoleAssignment.objects.get(
             sodar_uuid=SOURCE_CATEGORY_ROLE4_UUID
         )
-
         expected = {
             'id': c_owner_obj.pk,
             'project': category_obj.pk,
@@ -1070,24 +1103,17 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         }
         self.assertEqual(model_to_dict(c_guest_obj), expected)
 
+    def test_create_assert_project_roles(self):
+        """Test sync for creation and assert project role changes"""
+        self.assertEqual(RoleAssignment.objects.all().count(), 0)
+        self._update_remote_data()
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
+        self.assertEqual(RoleAssignment.objects.all().count(), 8)
+        new_user = User.objects.get(username=SOURCE_USER_USERNAME)
+        new_user2 = User.objects.get(username=SOURCE_USER2_USERNAME)
+        new_user3 = User.objects.get(username=SOURCE_USER3_USERNAME)
+        new_user4 = User.objects.get(username=SOURCE_USER4_USERNAME)
         project_obj = Project.objects.get(sodar_uuid=SOURCE_PROJECT_UUID)
-        expected = {
-            'id': project_obj.pk,
-            'title': SOURCE_PROJECT_TITLE,
-            'type': PROJECT_TYPE_PROJECT,
-            'description': SOURCE_PROJECT_DESCRIPTION,
-            'parent': category_obj.pk,
-            'public_access': None,
-            'public_guest_access': False,  # DEPRECATED
-            'archive': False,
-            'full_title': SOURCE_PROJECT_FULL_TITLE,
-            'has_public_children': False,
-            'sodar_uuid': uuid.UUID(SOURCE_PROJECT_UUID),
-        }
-        model_dict = model_to_dict(project_obj)
-        model_dict.pop('readme', None)
-        self.assertEqual(model_dict, expected)
-
         p_owner_obj = RoleAssignment.objects.get(
             sodar_uuid=SOURCE_PROJECT_ROLE_UUID
         )
@@ -1100,7 +1126,6 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         p_guest_obj = RoleAssignment.objects.get(
             sodar_uuid=SOURCE_PROJECT_ROLE4_UUID
         )
-
         expected = {
             'id': p_owner_obj.pk,
             'project': project_obj.pk,
@@ -1134,6 +1159,16 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         }
         self.assertEqual(model_to_dict(p_guest_obj), expected)
 
+    def test_create_assert_remote_objs(self):
+        """Test sync for creation and assert remote site and project"""
+        self.assertEqual(RemoteProject.objects.all().count(), 0)
+        self.assertEqual(RemoteSite.objects.all().count(), 1)
+        self._update_remote_data()
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
+        self.assertEqual(RemoteProject.objects.all().count(), 3)
+        self.assertEqual(RemoteSite.objects.all().count(), 2)
+        category_obj = Project.objects.get(sodar_uuid=SOURCE_CATEGORY_UUID)
+        project_obj = Project.objects.get(sodar_uuid=SOURCE_PROJECT_UUID)
         remote_cat_obj = RemoteProject.objects.get(
             site=self.source_site, project_uuid=category_obj.sodar_uuid
         )
@@ -1147,7 +1182,6 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
             'sodar_uuid': remote_cat_obj.sodar_uuid,
         }
         self.assertEqual(model_to_dict(remote_cat_obj), expected)
-
         remote_project_obj = RemoteProject.objects.get(
             site=self.source_site, project_uuid=project_obj.sodar_uuid
         )
@@ -1162,6 +1196,15 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         }
         self.assertEqual(model_to_dict(remote_project_obj), expected)
 
+    def test_create_assert_remote_peer_objs(self):
+        """Test sync for creation and assert remote peer objects"""
+        self.assertEqual(RemoteProject.objects.all().count(), 0)
+        self.assertEqual(RemoteSite.objects.all().count(), 1)
+        self._update_remote_data()
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
+        self.assertEqual(RemoteProject.objects.all().count(), 3)
+        self.assertEqual(RemoteSite.objects.all().count(), 2)
+        project_obj = Project.objects.get(sodar_uuid=SOURCE_PROJECT_UUID)
         peer_site_obj = RemoteSite.objects.get(
             sodar_uuid=PEER_SITE_UUID, mode=SITE_MODE_PEER
         )
@@ -1178,7 +1221,6 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         peer_site_dict = model_to_dict(peer_site_obj)
         peer_site_dict.pop('id')
         self.assertEqual(peer_site_dict, expected)
-
         peer_project_obj = RemoteProject.objects.get(site=peer_site_obj)
         expected = {
             'site': peer_site_obj.pk,
@@ -1192,7 +1234,14 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         peer_project_dict.pop('date_access')
         self.assertEqual(peer_project_dict, expected)
 
-        # Assert app settings
+    def test_create_assert_app_settings(self):
+        """Test sync for creation and assert app settings"""
+        self.assertEqual(AppSetting.objects.count(), 0)
+        self._update_remote_data()
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
+        self.assertEqual(AppSetting.objects.count(), 3)
+        new_user = User.objects.get(username=SOURCE_USER_USERNAME)
+        project_obj = Project.objects.get(sodar_uuid=SOURCE_PROJECT_UUID)
         expected = {
             'name': 'ip_restrict',
             'type': APP_SETTING_TYPE_BOOLEAN,
@@ -1227,8 +1276,13 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         }
         self.assert_app_setting(SET_STAR_UUID, expected)
 
-        # Assert remote_data changes
-        expected = og_data
+    def test_create_assert_remote_data(self):
+        """Test sync for creation and assert remote data changes"""
+        self.assertEqual(AppSetting.objects.count(), 0)
+        self._update_remote_data()
+        expected = deepcopy(self.remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
+        self.assertEqual(AppSetting.objects.count(), 3)
         expected['users'][SOURCE_USER_UUID]['status'] = 'created'
         expected['users'][SOURCE_USER2_UUID]['status'] = 'created'
         expected['users'][SOURCE_USER3_UUID]['status'] = 'created'
@@ -1262,15 +1316,16 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         expected['app_settings'][SET_IP_RESTRICT_UUID]['status'] = 'created'
         expected['app_settings'][SET_IP_ALLOW_LIST_UUID]['status'] = 'created'
         expected['app_settings'][SET_STAR_UUID]['status'] = 'created'
-        self.assertEqual(self.default_data, expected)
+        self.assertEqual(self.remote_data, expected)
 
     def test_create_app_setting_local(self):
         """Test sync with local app setting"""
-        remote_data = self.default_data
-        remote_data['app_settings'][SET_IP_RESTRICT_UUID]['global'] = False
-        remote_data['app_settings'][SET_IP_ALLOW_LIST_UUID]['global'] = False
-        expected = deepcopy(remote_data)
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        self.remote_data['app_settings'][SET_IP_RESTRICT_UUID]['global'] = False
+        self.remote_data['app_settings'][SET_IP_ALLOW_LIST_UUID][
+            'global'
+        ] = False
+        expected = deepcopy(self.remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
 
         expected['users'][SOURCE_USER_UUID]['status'] = 'created'
         expected['projects'][SOURCE_CATEGORY_UUID]['status'] = 'created'
@@ -1284,7 +1339,7 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         expected['app_settings'][SET_IP_RESTRICT_UUID]['status'] = 'created'
         expected['app_settings'][SET_IP_ALLOW_LIST_UUID]['status'] = 'created'
         expected['app_settings'][SET_STAR_UUID]['status'] = 'created'
-        self.assertEqual(remote_data, expected)
+        self.assertEqual(self.remote_data, expected)
 
     def test_create_multiple(self):
         """Test sync with multiple non-existing projects"""
@@ -1294,11 +1349,10 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         self.assertEqual(RemoteProject.objects.all().count(), 0)
         self.assertEqual(RemoteSite.objects.all().count(), 1)
 
-        remote_data = self.default_data
         new_project_uuid = str(uuid.uuid4())
         new_project_title = 'New Project Title'
         new_role_uuid = str(uuid.uuid4())
-        remote_data['projects'][new_project_uuid] = {
+        self.remote_data['projects'][new_project_uuid] = {
             'title': new_project_title,
             'type': PROJECT_TYPE_PROJECT,
             'level': REMOTE_LEVEL_READ_ROLES,
@@ -1312,8 +1366,8 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
                 }
             },
         }
-        og_data = deepcopy(remote_data)
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        ex_data = deepcopy(self.remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
 
         self.assertEqual(Project.objects.all().count(), 3)
         self.assertEqual(RoleAssignment.objects.all().count(), 3)
@@ -1351,24 +1405,23 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         }
         self.assertEqual(model_to_dict(p_new_owner_obj), expected)
 
-        expected = og_data
-        expected['users'][SOURCE_USER_UUID]['status'] = 'created'
-        expected['projects'][SOURCE_CATEGORY_UUID]['status'] = 'created'
-        expected['projects'][SOURCE_CATEGORY_UUID]['roles'][
+        ex_data['users'][SOURCE_USER_UUID]['status'] = 'created'
+        ex_data['projects'][SOURCE_CATEGORY_UUID]['status'] = 'created'
+        ex_data['projects'][SOURCE_CATEGORY_UUID]['roles'][
             SOURCE_CATEGORY_ROLE_UUID
         ]['status'] = 'created'
-        expected['projects'][SOURCE_PROJECT_UUID]['status'] = 'created'
-        expected['projects'][SOURCE_PROJECT_UUID]['roles'][
+        ex_data['projects'][SOURCE_PROJECT_UUID]['status'] = 'created'
+        ex_data['projects'][SOURCE_PROJECT_UUID]['roles'][
             SOURCE_PROJECT_ROLE_UUID
         ]['status'] = 'created'
-        expected['projects'][new_project_uuid]['status'] = 'created'
-        expected['projects'][new_project_uuid]['roles'][new_role_uuid][
+        ex_data['projects'][new_project_uuid]['status'] = 'created'
+        ex_data['projects'][new_project_uuid]['roles'][new_role_uuid][
             'status'
         ] = 'created'
-        expected['app_settings'][SET_IP_RESTRICT_UUID]['status'] = 'created'
-        expected['app_settings'][SET_IP_ALLOW_LIST_UUID]['status'] = 'created'
-        expected['app_settings'][SET_STAR_UUID]['status'] = 'created'
-        self.assertEqual(remote_data, expected)
+        ex_data['app_settings'][SET_IP_RESTRICT_UUID]['status'] = 'created'
+        ex_data['app_settings'][SET_IP_ALLOW_LIST_UUID]['status'] = 'created'
+        ex_data['app_settings'][SET_STAR_UUID]['status'] = 'created'
+        self.assertEqual(self.remote_data, ex_data)
 
     def test_create_local_owner(self):
         """Test sync with non-existing project data and local owner"""
@@ -1378,16 +1431,17 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         self.assertEqual(RemoteProject.objects.all().count(), 0)
         self.assertEqual(RemoteSite.objects.all().count(), 1)
 
-        remote_data = self.default_data
-        remote_data['users'][SOURCE_USER_UUID]['username'] = 'source_admin'
-        remote_data['users'][SOURCE_USER_UUID]['groups'] = [SYSTEM_USER_GROUP]
-        remote_data['projects'][SOURCE_CATEGORY_UUID]['roles'][
+        self.remote_data['users'][SOURCE_USER_UUID]['username'] = 'source_admin'
+        self.remote_data['users'][SOURCE_USER_UUID]['groups'] = [
+            SYSTEM_USER_GROUP
+        ]
+        self.remote_data['projects'][SOURCE_CATEGORY_UUID]['roles'][
             SOURCE_CATEGORY_ROLE_UUID
         ]['user'] = 'source_admin'
-        remote_data['projects'][SOURCE_PROJECT_UUID]['roles'][
+        self.remote_data['projects'][SOURCE_PROJECT_UUID]['roles'][
             SOURCE_PROJECT_ROLE_UUID
         ]['user'] = 'source_admin'
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
 
         self.assertEqual(Project.objects.all().count(), 2)
         self.assertEqual(RoleAssignment.objects.all().count(), 2)
@@ -1411,10 +1465,9 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         self.assertEqual(User.objects.all().count(), 1)
         self.assertEqual(RemoteProject.objects.all().count(), 0)
         self.assertEqual(RemoteSite.objects.all().count(), 1)
-        remote_data = self.default_data
         # Do sync, assert an exception is raised
         with self.assertRaises(ValueError):
-            self.remote_api.sync_remote_data(self.source_site, remote_data)
+            self.remote_api.sync_remote_data(self.source_site, self.remote_data)
 
         self.assertEqual(Project.objects.all().count(), 1)
         self.assertEqual(RoleAssignment.objects.all().count(), 0)
@@ -1428,8 +1481,7 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         new_user_username = 'newuser@' + SOURCE_USER_DOMAIN
         new_user_uuid = str(uuid.uuid4())
         new_role_uuid = str(uuid.uuid4())
-        remote_data = self.default_data
-        remote_data['users'][str(new_user_uuid)] = {
+        self.remote_data['users'][str(new_user_uuid)] = {
             'sodar_uuid': new_user_uuid,
             'username': new_user_username,
             'name': 'Some Name',
@@ -1439,10 +1491,10 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
             'additional_emails': [],
             'groups': [SOURCE_USER_GROUP],
         }
-        remote_data['projects'][SOURCE_CATEGORY_UUID]['roles'][
+        self.remote_data['projects'][SOURCE_CATEGORY_UUID]['roles'][
             new_role_uuid
         ] = {'user': new_user_username, 'role': PROJECT_ROLE_CONTRIBUTOR}
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
 
         self.assertEqual(User.objects.all().count(), 3)
         new_user = User.objects.get(username=new_user_username)
@@ -1454,15 +1506,14 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
 
     def test_create_no_access(self):
         """Test sync with no READ_ROLE access set"""
-        remote_data = self.default_data
-        remote_data['projects'][SOURCE_CATEGORY_UUID][
+        self.remote_data['projects'][SOURCE_CATEGORY_UUID][
             'level'
         ] = REMOTE_LEVEL_READ_INFO
-        remote_data['projects'][SOURCE_PROJECT_UUID][
+        self.remote_data['projects'][SOURCE_PROJECT_UUID][
             'level'
         ] = REMOTE_LEVEL_READ_INFO
-        og_data = deepcopy(remote_data)
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        expected = deepcopy(self.remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
 
         self.assertEqual(Project.objects.all().count(), 0)
         self.assertEqual(RoleAssignment.objects.all().count(), 0)
@@ -1470,15 +1521,14 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         self.assertEqual(RemoteProject.objects.all().count(), 0)
         self.assertEqual(RemoteSite.objects.all().count(), 1)
         # Assert no changes between update_data and remote_data
-        self.assertEqual(og_data, remote_data)
+        self.assertEqual(expected, self.remote_data)
 
     def test_create_local_user(self):
         """Test sync with local non-owner user"""
         local_user_username = 'localusername'
         local_user_uuid = str(uuid.uuid4())
         role_uuid = str(uuid.uuid4())
-        remote_data = self.default_data
-        remote_data['users'][local_user_uuid] = {
+        self.remote_data['users'][local_user_uuid] = {
             'sodar_uuid': local_user_uuid,
             'username': local_user_username,
             'name': SOURCE_USER_NAME,
@@ -1488,12 +1538,14 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
             'additional_emails': [],
             'groups': [SYSTEM_USER_GROUP],
         }
-        remote_data['projects'][SOURCE_PROJECT_UUID]['roles'][role_uuid] = {
+        self.remote_data['projects'][SOURCE_PROJECT_UUID]['roles'][
+            role_uuid
+        ] = {
             'user': local_user_username,
             'role': self.role_contributor.name,
         }
         self.assertEqual(User.objects.all().count(), 1)
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
 
         # Assert database status (the new user and role should not be created)
         self.assertEqual(Project.objects.all().count(), 2)
@@ -1508,9 +1560,8 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         local_user_username = 'localusername'
         local_user_uuid = str(uuid.uuid4())
         role_uuid = str(uuid.uuid4())
-        remote_data = self.default_data
         self.make_user(local_user_username)
-        remote_data['users'][local_user_uuid] = {
+        self.remote_data['users'][local_user_uuid] = {
             'sodar_uuid': local_user_uuid,
             'username': local_user_username,
             'name': SOURCE_USER_NAME,
@@ -1520,12 +1571,14 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
             'additional_emails': [],
             'groups': [SYSTEM_USER_GROUP],
         }
-        remote_data['projects'][SOURCE_PROJECT_UUID]['roles'][role_uuid] = {
+        self.remote_data['projects'][SOURCE_PROJECT_UUID]['roles'][
+            role_uuid
+        ] = {
             'user': local_user_username,
             'role': self.role_contributor.name,
         }
         self.assertEqual(User.objects.all().count(), 2)
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
 
         self.assertEqual(Project.objects.all().count(), 2)
         self.assertEqual(RoleAssignment.objects.all().count(), 3)
@@ -1539,8 +1592,7 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         local_user_username = 'localusername'
         local_user_uuid = str(uuid.uuid4())
         role_uuid = str(uuid.uuid4())
-        remote_data = self.default_data
-        remote_data['users'][local_user_uuid] = {
+        self.remote_data['users'][local_user_uuid] = {
             'sodar_uuid': local_user_uuid,
             'username': local_user_username,
             'name': SOURCE_USER_NAME,
@@ -1550,11 +1602,13 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
             'additional_emails': [],
             'groups': [SYSTEM_USER_GROUP],
         }
-        remote_data['projects'][SOURCE_PROJECT_UUID]['roles'][role_uuid] = {
+        self.remote_data['projects'][SOURCE_PROJECT_UUID]['roles'][
+            role_uuid
+        ] = {
             'user': local_user_username,
             'role': self.role_contributor.name,
         }
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
 
         self.assertEqual(Project.objects.all().count(), 2)
         self.assertEqual(RoleAssignment.objects.all().count(), 2)
@@ -1568,10 +1622,9 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         local_user_username = 'localusername'
         local_user_uuid = str(uuid.uuid4())
         role_uuid = str(uuid.uuid4())
-        remote_data = self.default_data
         # Create the user on the target site
         new_user = self.make_user(local_user_username)
-        remote_data['users'][local_user_uuid] = {
+        self.remote_data['users'][local_user_uuid] = {
             'sodar_uuid': local_user_uuid,
             'username': local_user_username,
             'name': SOURCE_USER_NAME,
@@ -1581,13 +1634,13 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
             'additional_emails': [],
             'groups': [SYSTEM_USER_GROUP],
         }
-        remote_data['projects'][SOURCE_PROJECT_UUID]['roles'] = {
+        self.remote_data['projects'][SOURCE_PROJECT_UUID]['roles'] = {
             role_uuid: {
                 'user': local_user_username,
                 'role': self.role_owner.name,
             }
         }
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
 
         self.assertEqual(Project.objects.all().count(), 2)
         self.assertEqual(RoleAssignment.objects.all().count(), 2)
@@ -1604,9 +1657,8 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         local_user_username = 'localusername'
         local_user_uuid = str(uuid.uuid4())
         role_uuid = str(uuid.uuid4())
-        remote_data = self.default_data
         # Create the user on the target site
-        remote_data['users'][local_user_uuid] = {
+        self.remote_data['users'][local_user_uuid] = {
             'sodar_uuid': local_user_uuid,
             'username': local_user_username,
             'name': SOURCE_USER_NAME,
@@ -1616,13 +1668,13 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
             'additional_emails': [],
             'groups': [SYSTEM_USER_GROUP],
         }
-        remote_data['projects'][SOURCE_PROJECT_UUID]['roles'] = {
+        self.remote_data['projects'][SOURCE_PROJECT_UUID]['roles'] = {
             role_uuid: {
                 'user': local_user_username,
                 'role': self.role_owner.name,
             }
         }
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
 
         self.assertEqual(Project.objects.all().count(), 2)
         self.assertEqual(RoleAssignment.objects.all().count(), 2)
@@ -1635,10 +1687,9 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
     def test_create_settings_user(self):
         """Test sync with USER scope settings"""
         self.assertEqual(AppSetting.objects.count(), 0)
-        remote_data = self.default_data
         global_set_uuid = str(uuid.uuid4())
         local_set_uuid = str(uuid.uuid4())
-        remote_data['app_settings'][global_set_uuid] = {
+        self.remote_data['app_settings'][global_set_uuid] = {
             'name': 'notify_email_project',
             'type': APP_SETTING_TYPE_BOOLEAN,
             'value': '0',
@@ -1649,7 +1700,7 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
             'user_name': SOURCE_USER_USERNAME,
             'global': True,
         }
-        remote_data['app_settings'][local_set_uuid] = {
+        self.remote_data['app_settings'][local_set_uuid] = {
             'name': 'user_str_setting',
             'type': APP_SETTING_TYPE_STRING,
             'value': 'Local value',
@@ -1661,7 +1712,7 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
             'global': False,
         }
 
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
         self.assertEqual(AppSetting.objects.count(), 5)
         target_user = User.objects.get(username=SOURCE_USER_USERNAME)
 
@@ -1701,20 +1752,21 @@ class TestSyncRemoteDataCreate(SyncRemoteDataTestBase):
         self.assertEqual(model_to_dict(obj), expected)
         # Assert sync data
         self.assertEqual(
-            remote_data['app_settings'][global_set_uuid]['status'], 'created'
+            self.remote_data['app_settings'][global_set_uuid]['status'],
+            'created',
         )
         self.assertEqual(
-            remote_data['app_settings'][local_set_uuid]['status'], 'created'
+            self.remote_data['app_settings'][local_set_uuid]['status'],
+            'created',
         )
 
     def test_create_user_add_email(self):
         """Test sync with additional email on user"""
         self.assertEqual(SODARUserAdditionalEmail.objects.count(), 0)
-        remote_data = self.default_data
-        remote_data['users'][SOURCE_USER_UUID]['additional_emails'] = [
+        self.remote_data['users'][SOURCE_USER_UUID]['additional_emails'] = [
             ADD_EMAIL
         ]
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
         self.assertEqual(SODARUserAdditionalEmail.objects.count(), 1)
         email = SODARUserAdditionalEmail.objects.first()
         target_user = User.objects.get(sodar_uuid=SOURCE_USER_UUID)
@@ -1728,6 +1780,43 @@ class TestSyncRemoteDataUpdate(
     SODARUserAdditionalEmailMixin, SyncRemoteDataTestBase
 ):
     """Tests for project updating with sync_remote_data()"""
+
+    def _update_remote_data(self):
+        """Update default data for remote update"""
+        # Add new user and contributor role to source project
+        self.remote_data['users'][str(self.new_user_uuid)] = {
+            'sodar_uuid': self.new_user_uuid,
+            'username': self.new_user_username,
+            'name': 'Some Name',
+            'first_name': 'Some',
+            'last_name': 'Name',
+            'email': 'some@example.com',
+            'additional_emails': [],
+            'groups': [SOURCE_USER_GROUP],
+        }
+        self.remote_data['projects'][SOURCE_PROJECT_UUID]['roles'][
+            self.new_role_uuid
+        ] = {
+            'user': self.new_user_username,
+            'role': PROJECT_ROLE_CONTRIBUTOR,
+        }
+
+        # Change Peer Site data
+        self.remote_data['peer_sites'][PEER_SITE_UUID]['name'] = NEW_PEER_NAME
+        self.remote_data['peer_sites'][PEER_SITE_UUID][
+            'description'
+        ] = NEW_PEER_DESC
+        self.remote_data['peer_sites'][PEER_SITE_UUID][
+            'user_display'
+        ] = NEW_PEER_USER_DISPLAY
+        # Change projectroles app settings
+        self.remote_data['app_settings'][SET_IP_RESTRICT_UUID]['value'] = True
+        self.remote_data['app_settings'][SET_IP_ALLOW_LIST_UUID][
+            'value'
+        ] = '192.168.1.1'
+        self.remote_data['projects'][SOURCE_CATEGORY_UUID]['status'] = 'updated'
+        self.remote_data['projects'][SOURCE_PROJECT_UUID]['status'] = 'updated'
+        self.remote_data['users'][SOURCE_USER_UUID]['status'] = 'updated'
 
     def setUp(self):
         super().setUp()
@@ -1748,7 +1837,6 @@ class TestSyncRemoteDataUpdate(
             readme='New readme',
             sodar_uuid=SOURCE_PROJECT_UUID,
         )
-
         # Set up user and roles
         self.user_target = self.make_sodar_user(
             username=SOURCE_USER_USERNAME,
@@ -1764,7 +1852,6 @@ class TestSyncRemoteDataUpdate(
         self.p_owner_obj = self.make_assignment(
             self.project_obj, self.user_target, self.role_owner
         )
-
         # Set up RemoteProject objects
         self.make_remote_project(
             project_uuid=self.category_obj.sodar_uuid,
@@ -1778,7 +1865,6 @@ class TestSyncRemoteDataUpdate(
             site=self.source_site,
             level=REMOTE_LEVEL_READ_ROLES,
         )
-
         # Set up Peer Objects
         self.peer_site = RemoteSite.objects.create(
             **{
@@ -1797,7 +1883,6 @@ class TestSyncRemoteDataUpdate(
             site=self.peer_site,
             level=SODAR_CONSTANTS['REMOTE_LEVEL_NONE'],
         )
-
         # Init app settings
         self.make_setting(
             plugin_name=APP_NAME,
@@ -1825,58 +1910,25 @@ class TestSyncRemoteDataUpdate(
             user=self.user_target,
             sodar_uuid=SET_STAR_UUID,
         )
+        # Set up IDs
+        self.new_user_username = 'newuser@' + SOURCE_USER_DOMAIN
+        self.new_user_uuid = str(uuid.uuid4())
+        self.new_role_uuid = str(uuid.uuid4())
+        # Update default data for all tests
+        self.remote_data['projects'][SOURCE_CATEGORY_UUID]['status'] = 'updated'
+        self.remote_data['projects'][SOURCE_PROJECT_UUID]['status'] = 'updated'
+        self.remote_data['users'][SOURCE_USER_UUID]['status'] = 'updated'
 
-        # Update default data
-        self.default_data['projects'][SOURCE_CATEGORY_UUID][
-            'status'
-        ] = 'updated'
-        self.default_data['projects'][SOURCE_PROJECT_UUID]['status'] = 'updated'
-        self.default_data['users'][SOURCE_USER_UUID]['status'] = 'updated'
-
-    def test_update(self):
-        """Test sync with existing project data and READ_ROLE access"""
+    def test_update_assert_db(self):
+        """Test sync for update and assert database changes"""
         self.assertEqual(Project.objects.all().count(), 2)
         self.assertEqual(RoleAssignment.objects.all().count(), 2)
         self.assertEqual(User.objects.all().count(), 2)
         self.assertEqual(RemoteProject.objects.all().count(), 3)
         self.assertEqual(RemoteSite.objects.all().count(), 2)
         self.assertEqual(AppSetting.objects.count(), 3)
-
-        remote_data = self.default_data
-        # Add new user and contributor role to source project
-        new_user_username = 'newuser@' + SOURCE_USER_DOMAIN
-        new_user_uuid = str(uuid.uuid4())
-        new_role_uuid = str(uuid.uuid4())
-        remote_data['users'][str(new_user_uuid)] = {
-            'sodar_uuid': new_user_uuid,
-            'username': new_user_username,
-            'name': 'Some Name',
-            'first_name': 'Some',
-            'last_name': 'Name',
-            'email': 'some@example.com',
-            'additional_emails': [],
-            'groups': [SOURCE_USER_GROUP],
-        }
-        remote_data['projects'][SOURCE_PROJECT_UUID]['roles'][new_role_uuid] = {
-            'user': new_user_username,
-            'role': PROJECT_ROLE_CONTRIBUTOR,
-        }
-
-        # Change Peer Site data
-        remote_data['peer_sites'][PEER_SITE_UUID]['name'] = NEW_PEER_NAME
-        remote_data['peer_sites'][PEER_SITE_UUID]['description'] = NEW_PEER_DESC
-        remote_data['peer_sites'][PEER_SITE_UUID][
-            'user_display'
-        ] = NEW_PEER_USER_DISPLAY
-        og_data = deepcopy(remote_data)
-        # Change projectroles app settings
-        remote_data['app_settings'][SET_IP_RESTRICT_UUID]['value'] = True
-        remote_data['app_settings'][SET_IP_ALLOW_LIST_UUID][
-            'value'
-        ] = '192.168.1.1'
-
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
-
+        self._update_remote_data()
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
         self.assertEqual(Project.objects.all().count(), 2)
         self.assertEqual(RoleAssignment.objects.all().count(), 3)
         self.assertEqual(User.objects.all().count(), 3)
@@ -1884,7 +1936,13 @@ class TestSyncRemoteDataUpdate(
         self.assertEqual(RemoteSite.objects.all().count(), 2)
         self.assertEqual(AppSetting.objects.count(), 3)
 
-        new_user = User.objects.get(username=new_user_username)
+    def test_update_assert_projects(self):
+        """Test sync for update and assert project changes"""
+        self.assertEqual(Project.objects.all().count(), 2)
+        self._update_remote_data()
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
+        self.assertEqual(Project.objects.all().count(), 2)
+
         self.category_obj.refresh_from_db()
         expected = {
             'id': self.category_obj.pk,
@@ -1903,16 +1961,6 @@ class TestSyncRemoteDataUpdate(
         model_dict.pop('readme', None)
         self.assertEqual(model_dict, expected)
 
-        self.c_owner_obj.refresh_from_db()
-        expected = {
-            'id': self.c_owner_obj.pk,
-            'project': self.category_obj.pk,
-            'user': self.user_target.pk,
-            'role': self.role_owner.pk,
-            'sodar_uuid': self.c_owner_obj.sodar_uuid,
-        }
-        self.assertEqual(model_to_dict(self.c_owner_obj), expected)
-
         self.project_obj.refresh_from_db()
         expected = {
             'id': self.project_obj.pk,
@@ -1930,6 +1978,26 @@ class TestSyncRemoteDataUpdate(
         model_dict = model_to_dict(self.project_obj)
         model_dict.pop('readme', None)
         self.assertEqual(model_dict, expected)
+
+    def test_update_assert_roles(self):
+        """Test sync for update and assert role changes"""
+        self.assertEqual(RoleAssignment.objects.all().count(), 2)
+        self.assertEqual(User.objects.all().count(), 2)
+        self._update_remote_data()
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
+        self.assertEqual(RoleAssignment.objects.all().count(), 3)
+        self.assertEqual(User.objects.all().count(), 3)
+        new_user = User.objects.get(username=self.new_user_username)
+
+        self.c_owner_obj.refresh_from_db()
+        expected = {
+            'id': self.c_owner_obj.pk,
+            'project': self.category_obj.pk,
+            'user': self.user_target.pk,
+            'role': self.role_owner.pk,
+            'sodar_uuid': self.c_owner_obj.sodar_uuid,
+        }
+        self.assertEqual(model_to_dict(self.c_owner_obj), expected)
 
         self.p_owner_obj.refresh_from_db()
         expected = {
@@ -1953,6 +2021,17 @@ class TestSyncRemoteDataUpdate(
             'sodar_uuid': p_contrib_obj.sodar_uuid,
         }
         self.assertEqual(model_to_dict(p_contrib_obj), expected)
+
+    def test_update_assert_remote_objs(self):
+        """Test sync for update and assert remote site and project"""
+        self.assertEqual(RemoteProject.objects.all().count(), 3)
+        self.assertEqual(RemoteSite.objects.all().count(), 2)
+        self._update_remote_data()
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
+        self.assertEqual(RemoteProject.objects.all().count(), 3)
+        self.assertEqual(RemoteSite.objects.all().count(), 2)
+        self.category_obj.refresh_from_db()
+        self.project_obj.refresh_from_db()
 
         remote_cat_obj = RemoteProject.objects.get(
             site=self.source_site, project_uuid=self.category_obj.sodar_uuid
@@ -1981,6 +2060,17 @@ class TestSyncRemoteDataUpdate(
             'sodar_uuid': remote_project_obj.sodar_uuid,
         }
         self.assertEqual(model_to_dict(remote_project_obj), expected)
+
+    def test_update_assert_remote_peer_objs(self):
+        """Test sync for updating and assert remote peer objects"""
+        self.assertEqual(RemoteProject.objects.all().count(), 3)
+        self.assertEqual(RemoteSite.objects.all().count(), 2)
+        self._update_remote_data()
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
+        self.assertEqual(RemoteProject.objects.all().count(), 3)
+        self.assertEqual(RemoteSite.objects.all().count(), 2)
+        self.category_obj.refresh_from_db()
+        self.project_obj.refresh_from_db()
 
         peer_site_obj = RemoteSite.objects.get(
             sodar_uuid=PEER_SITE_UUID, mode=SITE_MODE_PEER
@@ -2012,7 +2102,12 @@ class TestSyncRemoteDataUpdate(
         peer_project_dict.pop('date_access')
         self.assertEqual(peer_project_dict, expected)
 
-        # Assert app settings
+    def test_update_assert_app_settings(self):
+        """Test sync for updating and assert app settings"""
+        self.assertEqual(AppSetting.objects.count(), 3)
+        self._update_remote_data()
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
+        self.assertEqual(AppSetting.objects.count(), 3)
         # NOTE: Global app settings should not be updated
         expected = {
             'name': 'ip_restrict',
@@ -2048,13 +2143,16 @@ class TestSyncRemoteDataUpdate(
         }
         self.assert_app_setting(SET_STAR_UUID, expected)
 
-        # Assert update_data changes
-        expected = og_data
+    def test_update_assert_remote_data(self):
+        """Test sync for updating and assert remote data changes"""
+        self._update_remote_data()
+        expected = deepcopy(self.remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
         expected['users'][SOURCE_USER_UUID]['status'] = 'updated'
-        expected['users'][new_user_uuid]['status'] = 'created'
+        expected['users'][self.new_user_uuid]['status'] = 'created'
         expected['projects'][SOURCE_CATEGORY_UUID]['status'] = 'updated'
         expected['projects'][SOURCE_PROJECT_UUID]['status'] = 'updated'
-        expected['projects'][SOURCE_PROJECT_UUID]['roles'][new_role_uuid][
+        expected['projects'][SOURCE_PROJECT_UUID]['roles'][self.new_role_uuid][
             'status'
         ] = 'created'
         expected['app_settings'][SET_IP_RESTRICT_UUID]['value'] = True
@@ -2064,7 +2162,7 @@ class TestSyncRemoteDataUpdate(
         expected['app_settings'][SET_IP_RESTRICT_UUID]['status'] = 'updated'
         expected['app_settings'][SET_IP_ALLOW_LIST_UUID]['status'] = 'updated'
         expected['app_settings'][SET_STAR_UUID]['status'] = 'skipped'
-        self.assertEqual(remote_data, expected)
+        self.assertEqual(self.remote_data, expected)
 
     def test_update_user_uuid_ldap(self):
         """Test sync with legacy UUID for LDAP user"""
@@ -2074,7 +2172,7 @@ class TestSyncRemoteDataUpdate(
         self.user_target.sodar_uuid = target_uuid
         self.user_target.save()
         self.assertEqual(User.objects.all().count(), 2)
-        self.remote_api.sync_remote_data(self.source_site, self.default_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
         self.assertEqual(User.objects.all().count(), 2)
         self.user_target.refresh_from_db()
         self.assertEqual(str(self.user_target.sodar_uuid), SOURCE_USER_UUID)
@@ -2089,8 +2187,7 @@ class TestSyncRemoteDataUpdate(
         local_uuid_source = str(uuid.uuid4())
         self.assertNotEqual(local_uuid_target, local_uuid_source)
         self.assertNotEqual(user_local.name, local_name)
-        remote_data = self.default_data
-        remote_data['users'][local_uuid_source] = {
+        self.remote_data['users'][local_uuid_source] = {
             'username': local_username,
             'name': local_name,
             'first_name': local_name.split(' ')[0],
@@ -2100,7 +2197,7 @@ class TestSyncRemoteDataUpdate(
             'groups': [SOURCE_USER_GROUP],
         }
         self.assertEqual(User.objects.all().count(), 3)
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
         self.assertEqual(User.objects.all().count(), 3)
         user_local.refresh_from_db()
         self.assertEqual(str(user_local.sodar_uuid), local_uuid_source)
@@ -2118,8 +2215,7 @@ class TestSyncRemoteDataUpdate(
         local_uuid_source = str(uuid.uuid4())  # Source UUID
         self.assertNotEqual(local_uuid_target, local_uuid_source)
         self.assertNotEqual(user_local.name, local_name)
-        remote_data = self.default_data
-        remote_data['users'][local_uuid_source] = {
+        self.remote_data['users'][local_uuid_source] = {
             'username': local_username,
             'name': local_name,
             'first_name': local_name.split(' ')[0],
@@ -2129,7 +2225,7 @@ class TestSyncRemoteDataUpdate(
             'groups': [SYSTEM_USER_GROUP],
         }
         self.assertEqual(User.objects.all().count(), 3)
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
         self.assertEqual(User.objects.all().count(), 3)
         user_local.refresh_from_db()
         self.assertEqual(str(user_local.sodar_uuid), local_uuid_target)
@@ -2138,12 +2234,11 @@ class TestSyncRemoteDataUpdate(
     def test_update_inherited(self):
         """Test sync with existing project data and inherited contributor"""
         self.assertEqual(User.objects.all().count(), 2)
-        remote_data = self.default_data
         # Add new user and contributor role to source category
         new_user_username = 'newuser@' + SOURCE_USER_DOMAIN
         new_user_uuid = str(uuid.uuid4())
         new_role_uuid = str(uuid.uuid4())
-        remote_data['users'][str(new_user_uuid)] = {
+        self.remote_data['users'][str(new_user_uuid)] = {
             'sodar_uuid': new_user_uuid,
             'username': new_user_username,
             'name': 'Some Name',
@@ -2153,10 +2248,10 @@ class TestSyncRemoteDataUpdate(
             'additional_emails': [],
             'groups': [SOURCE_USER_GROUP],
         }
-        remote_data['projects'][SOURCE_CATEGORY_UUID]['roles'][
+        self.remote_data['projects'][SOURCE_CATEGORY_UUID]['roles'][
             new_role_uuid
         ] = {'user': new_user_username, 'role': PROJECT_ROLE_CONTRIBUTOR}
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
 
         self.assertEqual(User.objects.all().count(), 3)
         new_user = User.objects.get(username=new_user_username)
@@ -2169,15 +2264,16 @@ class TestSyncRemoteDataUpdate(
 
     def test_update_settings_local(self):
         """Test update with local app settings (should not be updated)"""
-        remote_data = self.default_data
-        remote_data['app_settings'][SET_IP_RESTRICT_UUID]['global'] = False
-        remote_data['app_settings'][SET_IP_ALLOW_LIST_UUID]['global'] = False
-        remote_data['app_settings'][SET_IP_RESTRICT_UUID]['value'] = True
-        remote_data['app_settings'][SET_IP_ALLOW_LIST_UUID]['value_json'] = [
-            '192.168.1.1'
-        ]
-        og_data = deepcopy(remote_data)
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        self.remote_data['app_settings'][SET_IP_RESTRICT_UUID]['global'] = False
+        self.remote_data['app_settings'][SET_IP_ALLOW_LIST_UUID][
+            'global'
+        ] = False
+        self.remote_data['app_settings'][SET_IP_RESTRICT_UUID]['value'] = True
+        self.remote_data['app_settings'][SET_IP_ALLOW_LIST_UUID][
+            'value_json'
+        ] = ['192.168.1.1']
+        ex_data = deepcopy(self.remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
 
         expected = {
             'name': 'ip_restrict',
@@ -2213,22 +2309,21 @@ class TestSyncRemoteDataUpdate(
         }
         self.assert_app_setting(SET_STAR_UUID, expected)
 
-        og_data['users'][SOURCE_USER_UUID]['status'] = 'updated'
-        og_data['projects'][SOURCE_CATEGORY_UUID]['status'] = 'updated'
-        og_data['projects'][SOURCE_PROJECT_UUID]['status'] = 'updated'
-        og_data['app_settings'][SET_IP_RESTRICT_UUID]['status'] = 'skipped'
-        og_data['app_settings'][SET_IP_ALLOW_LIST_UUID]['status'] = 'skipped'
-        og_data['app_settings'][SET_STAR_UUID]['status'] = 'skipped'
-        self.assertEqual(remote_data, og_data)
+        ex_data['users'][SOURCE_USER_UUID]['status'] = 'updated'
+        ex_data['projects'][SOURCE_CATEGORY_UUID]['status'] = 'updated'
+        ex_data['projects'][SOURCE_PROJECT_UUID]['status'] = 'updated'
+        ex_data['app_settings'][SET_IP_RESTRICT_UUID]['status'] = 'skipped'
+        ex_data['app_settings'][SET_IP_ALLOW_LIST_UUID]['status'] = 'skipped'
+        ex_data['app_settings'][SET_STAR_UUID]['status'] = 'skipped'
+        self.assertEqual(self.remote_data, ex_data)
 
     def test_update_settings_no_app(self):
         """Test update with app setting for app not present on target site"""
         self.assertEqual(AppSetting.objects.count(), 3)
-        remote_data = self.default_data
         setting_uuid = str(uuid.uuid4())
         setting_name = 'NOT_A_VALID_SETTING'
         # Change projectroles app settings
-        remote_data['app_settings'][setting_uuid] = {
+        self.remote_data['app_settings'][setting_uuid] = {
             'name': setting_name,
             'type': APP_SETTING_TYPE_BOOLEAN,
             'value': False,
@@ -2238,7 +2333,7 @@ class TestSyncRemoteDataUpdate(
             'user_uuid': None,
             'global': True,
         }
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
         # Make sure setting was not set
         self.assertIsNone(AppSetting.objects.filter(name=setting_name).first())
 
@@ -2261,10 +2356,9 @@ class TestSyncRemoteDataUpdate(
         )
         self.assertEqual(AppSetting.objects.count(), 5)
 
-        remote_data = self.default_data
         global_set_uuid = str(uuid.uuid4())
         local_set_uuid = str(uuid.uuid4())
-        remote_data['app_settings'][global_set_uuid] = {
+        self.remote_data['app_settings'][global_set_uuid] = {
             'name': 'notify_email_project',
             'type': APP_SETTING_TYPE_BOOLEAN,
             'value': '0',
@@ -2275,7 +2369,7 @@ class TestSyncRemoteDataUpdate(
             'user_name': SOURCE_USER_USERNAME,
             'global': True,
         }
-        remote_data['app_settings'][local_set_uuid] = {
+        self.remote_data['app_settings'][local_set_uuid] = {
             'name': 'user_str_setting',
             'type': APP_SETTING_TYPE_STRING,
             'value': 'Source value',
@@ -2286,7 +2380,7 @@ class TestSyncRemoteDataUpdate(
             'user_name': SOURCE_USER_USERNAME,
             'global': False,
         }
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
 
         self.assertEqual(AppSetting.objects.count(), 5)
         # Global setting should be updated
@@ -2297,10 +2391,12 @@ class TestSyncRemoteDataUpdate(
         self.assertEqual(target_local_setting.value, 'Target value')
         # Assert sync data
         self.assertEqual(
-            remote_data['app_settings'][global_set_uuid]['status'], 'updated'
+            self.remote_data['app_settings'][global_set_uuid]['status'],
+            'updated',
         )
         self.assertEqual(
-            remote_data['app_settings'][local_set_uuid]['status'], 'skipped'
+            self.remote_data['app_settings'][local_set_uuid]['status'],
+            'skipped',
         )
 
     def test_update_revoke(self):
@@ -2322,13 +2418,12 @@ class TestSyncRemoteDataUpdate(
         self.assertEqual(RemoteProject.objects.all().count(), 3)
         self.assertEqual(RemoteSite.objects.all().count(), 2)
 
-        remote_data = self.default_data
         # Revoke access to project
-        remote_data['projects'][SOURCE_PROJECT_UUID][
+        self.remote_data['projects'][SOURCE_PROJECT_UUID][
             'level'
         ] = REMOTE_LEVEL_REVOKED
-        remote_data['projects'][SOURCE_PROJECT_UUID]['remote_sites'] = []
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        self.remote_data['projects'][SOURCE_PROJECT_UUID]['remote_sites'] = []
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
 
         self.assertEqual(Project.objects.all().count(), 2)
         self.assertEqual(RoleAssignment.objects.all().count(), 2)
@@ -2344,10 +2439,12 @@ class TestSyncRemoteDataUpdate(
                 role__name=PROJECT_ROLE_CONTRIBUTOR,
             )
         self.assertEqual(
-            remote_data['projects'][SOURCE_PROJECT_UUID]['level'],
+            self.remote_data['projects'][SOURCE_PROJECT_UUID]['level'],
             REMOTE_LEVEL_REVOKED,
         )
-        self.assertNotIn(str(new_user.sodar_uuid), remote_data['users'].keys())
+        self.assertNotIn(
+            str(new_user.sodar_uuid), self.remote_data['users'].keys()
+        )
 
     def test_delete_role(self):
         """Test sync with existing project data and removed role"""
@@ -2365,9 +2462,8 @@ class TestSyncRemoteDataUpdate(
         self.assertEqual(RemoteProject.objects.all().count(), 3)
         self.assertEqual(RemoteSite.objects.all().count(), 2)
 
-        remote_data = self.default_data
-        og_data = deepcopy(remote_data)
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        expected = deepcopy(self.remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
 
         self.assertEqual(Project.objects.all().count(), 2)
         self.assertEqual(RoleAssignment.objects.all().count(), 2)
@@ -2380,15 +2476,15 @@ class TestSyncRemoteDataUpdate(
                 role__name=PROJECT_ROLE_CONTRIBUTOR,
             )
 
-        og_data['projects'][SOURCE_PROJECT_UUID]['roles'][new_role_uuid] = {
+        expected['projects'][SOURCE_PROJECT_UUID]['roles'][new_role_uuid] = {
             'user': new_user_username,
             'role': PROJECT_ROLE_CONTRIBUTOR,
             'status': 'deleted',
         }
-        og_data['app_settings'][SET_IP_RESTRICT_UUID]['status'] = 'skipped'
-        og_data['app_settings'][SET_IP_ALLOW_LIST_UUID]['status'] = 'skipped'
-        og_data['app_settings'][SET_STAR_UUID]['status'] = 'skipped'
-        self.assertEqual(remote_data, og_data)
+        expected['app_settings'][SET_IP_RESTRICT_UUID]['status'] = 'skipped'
+        expected['app_settings'][SET_IP_ALLOW_LIST_UUID]['status'] = 'skipped'
+        expected['app_settings'][SET_STAR_UUID]['status'] = 'skipped'
+        self.assertEqual(self.remote_data, expected)
 
     def test_update_no_changes(self):
         """Test sync with existing project data and no changes"""
@@ -2398,9 +2494,8 @@ class TestSyncRemoteDataUpdate(
         self.assertEqual(RemoteProject.objects.all().count(), 3)
         self.assertEqual(RemoteSite.objects.all().count(), 2)
 
-        remote_data = self.default_data
-        og_data = deepcopy(remote_data)
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        ex_data = deepcopy(self.remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
 
         self.assertEqual(Project.objects.all().count(), 2)
         self.assertEqual(RoleAssignment.objects.all().count(), 2)
@@ -2522,19 +2617,18 @@ class TestSyncRemoteDataUpdate(
         peer_project_dict.pop('date_access')
         self.assertEqual(peer_project_dict, expected)
 
-        og_data['app_settings'][SET_IP_RESTRICT_UUID]['status'] = 'skipped'
-        og_data['app_settings'][SET_IP_ALLOW_LIST_UUID]['status'] = 'skipped'
-        og_data['app_settings'][SET_STAR_UUID]['status'] = 'skipped'
-        self.assertEqual(og_data, remote_data)
+        ex_data['app_settings'][SET_IP_RESTRICT_UUID]['status'] = 'skipped'
+        ex_data['app_settings'][SET_IP_ALLOW_LIST_UUID]['status'] = 'skipped'
+        ex_data['app_settings'][SET_STAR_UUID]['status'] = 'skipped'
+        self.assertEqual(ex_data, self.remote_data)
 
     def test_update_user_add_email(self):
         """Test sync with new additional email on user"""
         self.assertEqual(SODARUserAdditionalEmail.objects.count(), 0)
-        remote_data = self.default_data
-        remote_data['users'][SOURCE_USER_UUID]['additional_emails'] = [
+        self.remote_data['users'][SOURCE_USER_UUID]['additional_emails'] = [
             ADD_EMAIL
         ]
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
         self.assertEqual(SODARUserAdditionalEmail.objects.count(), 1)
         email = SODARUserAdditionalEmail.objects.first()
         target_user = User.objects.get(sodar_uuid=SOURCE_USER_UUID)
@@ -2552,18 +2646,16 @@ class TestSyncRemoteDataUpdate(
             secret=build_secret(16),
         )
         self.assertEqual(SODARUserAdditionalEmail.objects.count(), 1)
-        remote_data = self.default_data
-        remote_data['users'][SOURCE_USER_UUID]['additional_emails'] = [
+        self.remote_data['users'][SOURCE_USER_UUID]['additional_emails'] = [
             ADD_EMAIL
         ]
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
         self.assertEqual(SODARUserAdditionalEmail.objects.count(), 1)
 
     def test_update_user_add_email_delete(self):
         """Test sync with deleting existing additional email on user"""
         self.make_email(self.user_target, ADD_EMAIL)
         self.assertEqual(SODARUserAdditionalEmail.objects.count(), 1)
-        remote_data = self.default_data
-        remote_data['users'][SOURCE_USER_UUID]['additional_emails'] = []
-        self.remote_api.sync_remote_data(self.source_site, remote_data)
+        self.remote_data['users'][SOURCE_USER_UUID]['additional_emails'] = []
+        self.remote_api.sync_remote_data(self.source_site, self.remote_data)
         self.assertEqual(SODARUserAdditionalEmail.objects.count(), 0)